requirements.txt
together
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
google-auth
#google-auth-oauthlib-tool
python-dotenv
requests
openai
elevenlabs
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
types-requests>=2.31.0.20240218
together>=0.2.8
google-generativeai>=0.3.2
pydantic>=2.6.1
moviepy>=1.0.3
openai-whisper>=20231117
numpy<2.0.0
torch>=2.1.2
whisper>=1.1.10
soundfile>=0.12.1
librosa>=0.10.1
pydub>=0.25.1
ffmpeg-python>=0.2.0
pytest
pandas